supervisor/requirements.txt
langchain
langgraph
langsmith
kafka-python
requests
mcp
